server/seesun/requirements.txt
loguru==0.7.2
pydantic-settings==2.6.1
psycopg2-binary==2.9.9
requests==2.32.3

